Added seperate loading of TemplateScreen.py data and RelationChangeScreen.py
now when you open a project you come to the TemplateScreen.py and it only loads in the subset and the classes in the subset. Only when you go to the RelationChangeScreen.py the RelationChangeDomain.py starts loading the saved data of the project with all the lists showing a message that the data is being loaded.