Work around for DataVisualisationScreen.py when installing with Admin rights.
When installing with admin rights it gets installed in Program Files folder and by defual the pyvis library will try to create and write to the home directory but it can't when the app doesn't have admin rights. By default it doesn't and you shouldn't let it have admin rights.

This work arround will change the current working directory of the python to %user_home%/OLTWizardProjects/, where the pyvis library is allowed to write, everytime the PyVisWrapper is called after the call the original current working directory is restored